requirements.txt
# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# HTTP 请求
requests>=2.31.0
urllib3>=2.0.0

# 技术分析
ta-lib>=0.4.28
TA-Lib>=0.4.28

# LLM API
openai>=1.12.0

# 配置文件处理
pyyaml>=6.0

# RSS 解析
feedparser>=6.0.10

# 日志
loguru>=0.7.0

# 数据可视化
matplotlib>=3.7.0
plotly>=5.18.0

# 日期时间
python-dateutil>=2.8.2

# HTML 解析
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 环境变量
python-dotenv>=1.0.0

# 进度条
tqdm>=4.65.0
